modified front end as per wireframe

+++ src/components/LessonTabItem.js
@@ -11,7 +11,10 @@ export default class LessonTabItem extends React.Component {
         return (
             <div>{this.props.lesson.title}&nbsp;
                 <i id='deleteLessonBtn' className='fa fa-times' onClick={() => {
-                    this.props.deleteLesson(this.props.lesson.id)
+                    if (window.confirm('Are you sure you want to delete lesson?')) {
+                        this.props.deleteLesson(this.props.lesson.id);
+                    }
+
                 }}>
                 </i>
             </div>
